#71 - Populate repeating block groups from file
Even though the group_class is being passed the index, it is not
being used at the point level, and any update that I tried to use
it at the point level resulted in the output changing, and lots of tests
failing.

Using the index as part of the _init_repeating_group instead, without
passing it to the group_class, provides the expected results.

Note that this follows the same logic as if the count is available (line
530).